Sandboxed incompatible tests
